@@ -15,7 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.xml.bind.ValidationException;
+import jakarta.validation.ValidationException;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashMap;
