@@ -9,7 +9,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.xml.bind.ValidationException;
+import jakarta.validation.ValidationException;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.HashMap;
