@@ -6,7 +6,7 @@ import com.therearenotasksforus.videohostingapi.models.Profile;
 import com.therearenotasksforus.videohostingapi.models.User;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.xml.bind.ValidationException;
+import jakarta.validation.ValidationException;
 import java.util.List;
 
 public interface ProfileService {
