@@ -29,16 +29,13 @@ public class FileStore {
                      Optional<Map<String, String>> optionalMetadata,
                      InputStream inputStream) {
         ObjectMetadata metadata = new ObjectMetadata();
-
         optionalMetadata.ifPresent(map -> {
             if (!map.isEmpty()) {
                 map.forEach(metadata::addUserMetadata);
             }
         });
-
         try {
-            s3.putObject(new PutObjectRequest(path, fileName, inputStream, metadata)
-                    .withCannedAcl(CannedAccessControlList.PublicRead));
+            s3.putObject(new PutObjectRequest(path, fileName, inputStream, metadata));
         } catch (AmazonServiceException e) {
             throw new IllegalStateException("Failed to store file to s3", e);
         }
