@@ -1,13 +1,9 @@
 package com.therearenotasksforus.videohostingapi.models;
 
+import jakarta.persistence.*;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
-import javax.persistence.Id;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import java.sql.Timestamp;
 
 @MappedSuperclass
